requirements.txt
django
djangorestframework
django-mysql